Add atividade complementar 1

+++ modulo3/apis-labenusers/apis-labenusers/src/Components/UsersList.js
@@ -6,17 +6,22 @@ export function UsersList (props) {
     const renderList = props.usersListProps.map(item => {
 
         const delUser = () =>{
-            axios.delete(props.urlProp + "/" + item.id, props.headersProp).then((response) =>{
-                alert('User vaporized.')
-            }).catch((error) =>{
+            if (window.confirm('Eliminate user?')) {
+                axios.delete(props.urlProp + "/" + item.id, props.headersProp).then((response) =>{
+                    alert('User vaporized.')
+                }).catch((error) =>{
+                    alert("User survived.")
+                })
+            }else {
                 alert("User survived.")
-            })
+            }
         }
 
         return (
             <>
                 <All.Span>
                     <li>{item.name}</li>
+                    {/* <button onClick={() => delUser}>x</button> */}
                     <button onClick={delUser}>x</button>
                 </All.Span>
